style(gemma4): drop redundant i32 cast tripping the PP-modules clippy gate (#688)
The pipeline-parallel CI's `clippy -p mlxcel --lib --tests -D warnings` step is path-gated to PRs touching the distributed modules, so this pre-existing unnecessary_cast in gemma4_tests.rs (from #679) only surfaced now that this branch adds the pipeline-parallel guard. Trivial lint fix, no behavior change.

+++ src/models/gemma4_tests.rs
@@ -273,7 +273,7 @@ mod cache_isolation {
                 let weights = make_test_gemma4_weight_map();
                 Gemma4Wrapper::new(Gemma4Model::from_weights(&weights, &args).unwrap())
             };
-            let prompt: Vec<i32> = (0..24).map(|i| (i % 7) as i32).collect();
+            let prompt: Vec<i32> = (0..24).map(|i| i % 7).collect();
 
             let single = build();
             let mut caches = single.make_caches();
